+++ qiskit/circuit/__init__.py
@@ -230,3 +230,4 @@ from .parameter import Parameter
 from .parametervector import ParameterVector
 from .parameterexpression import ParameterExpression
 from .equivalence import EquivalenceLibrary
+from .classicalfunction.types import Int1, Int2
