+++ qiskit/circuit/classicalfunction/compile_classical_function.py
@@ -1,0 +1,25 @@
+# This code is part of Qiskit.
+#
+# (C) Copyright IBM 2020.
+#
+# This code is licensed under the Apache License, Version 2.0. You may
+# obtain a copy of this license in the LICENSE.txt file in the root directory
+# of this source tree or at http://www.apache.org/licenses/LICENSE-2.0.
+#
+# Any modifications or derivative works of this code must retain this
+# copyright notice, and modified files need to carry a notice indicating
+# that they have been altered from the originals.
+
+"""=============================================
+ClassicalFunction compiler (:mod:`qiskit.circuit.classicalfunction.compile_classical_function`)
+=============================================
+
+.. currentmodule:: qiskit.circuit.classicalfunction.compile_classical_function
+
+.. autofunction:: execute
+"""
+
+import inspect
+from .classicalfunction import ClassicalFunction
+
+
