+++ qiskit/circuit/classicalfunction/types.py
