+++ test/python/classical_function_compiler/test_classical_function.py
@@ -14,7 +14,7 @@
 
 from qiskit.test import QiskitTestCase
 
-from qiskit.circuit.classical_function import classical_function as compile_classical_function
+from qiskit.circuit.classicalfunction import classical_function as compile_classical_function
 
 from qiskit import QuantumCircuit
 from qiskit.circuit.library.standard_gates import XGate
