+++ test/python/classical_function_compiler/utils.py
@@ -10,14 +10,14 @@
 # copyright notice, and modified files need to carry a notice indicating
 # that they have been altered from the originals.
 
-"""Testing utilities for classical_function compiler."""
+"""Testing utilities for classicalfunction compiler."""
 
 from inspect import getfullargspec, isfunction
 from . import examples
 
 
 def get_truthtable_from_function(function):
-    """Runs an classical_function function (on python) and generates a truthtable from it."""
+    """Runs an classicalfunction function (on python) and generates a truthtable from it."""
     amount_bit_input = len(getfullargspec(function).args)
     result = ""
     for decimal in range(2 ** amount_bit_input):
